rename user operation validation to sanity check, move to separate file, prepare for user operation simulation
