+++ src/uopool/services/mod.rs
@@ -1,3 +1,5 @@
+mod sanity_check;
+mod simulation;
 mod uopool;
 
 pub use self::uopool::*;
